requirements.txt
beautifulsoup4==4.12.0
selenium==3.141.0
requests==2.26.0
python-dotenv==0.21.1

